Fix import statements after moving extension init to extensions.py
