+++ app/db/models.py
@@ -1,4 +1,4 @@
-from app.db import db
+from app.extensions import db
 from datetime import datetime
